refactor: move decks (page) to decks-page
