+++ src/pages/decks-page/flashcard-decks.scss
